services/auth/authentication/pyproject.toml
[project]
name = "authentication"
version = "0.0.1"
description = "Service:"
authors = [{ name = "Dmitry", email = "dbevzpr@mail.ru" }]
readme = "README.md"
keywords = ['python']
requires-python = ">=3.12"
classifiers = [
    "Intended Audience :: Developers",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "pydantic-settings>=2.7.1",
    "python-dotenv>=1.0.1",
    "sqlmodel>=0.0.22",
    "ydb-sqlalchemy>=0.1.6",
    "ydb[yc]>=3.18.15",
]


[dependency-groups]
dev = [
    "uvicorn>=0.34.0",
    "pytest>=7.2.0",
    "pre-commit>=2.20.0",
    "mypy>=0.991",
    "ruff>=0.9.2",
]

[tool.mypy]
files = [
    "authentication"
]
disallow_untyped_defs = true
disallow_any_unimported = true
no_implicit_optional = true
check_untyped_defs = true
warn_return_any = true
warn_unused_ignores = true
show_error_codes = true

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py39"
line-length = 120
fix = true

[tool.ruff.lint]
select = [
    # flake8-2020
    "YTT",
    # flake8-bandit
    "S",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # flake8-comprehensions
    "C4",
    # flake8-debugger
    "T10",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # mccabe
    "C90",
    # pycodestyle
    "E", "W",
    # pyflakes
    "F",
    # pygrep-hooks
    "PGH",
    # pyupgrade
    "UP",
    # ruff
    "RUF",
    # tryceratops
    "TRY",
]
ignore = [
    # LineTooLong
    "E501",
    # DoNotAssignLambda
    "E731",
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]

[tool.ruff.format]
preview = true


[tool.pytest]
testpaths = ["tests"]
